moving files around + page setup
