+++ frontend/webapp/src/app/Backlog/page.js
@@ -1,0 +1,10 @@
+import React from "react";
+import Navbar from "../components/Navbar";
+
+export default function Page(){
+    return (
+        <div className="w-screen h-screen backgroundImg">
+            <Navbar/>
+        </div>
+    )
+}
