+++ frontend/webapp/src/app/components/BotStatus.js
