+++ frontend/webapp/src/app/page.js
@@ -2,7 +2,7 @@ import React from "react";
 import ChartCarosel from "./components/chart-carosels";
 import Table from "./components/Table";
 import Navbar from "./components/Navbar";
-import BotStatus from "./BotStatus";
+import BotStatus from "./components/BotStatus";
 
 export default function Home() {
   return (
